+++ src/app/about/about.component.html
@@ -20,9 +20,6 @@
         <a mat-button href="https://www.paypal.me/martz">Donate</a>
       </mat-list-item>
     </mat-list>
-    <ul class="nav flex-column">
-
-    </ul>
   </mat-sidenav>
   <mat-sidenav-content>
     <mat-card>
