+++ src/app/courses/course-list/course-list.component.html
@@ -1,0 +1,95 @@
+<mat-card>
+    <mat-card-title *ngIf="!shortened">
+      Courses
+    </mat-card-title>
+    <mat-card-subtitle *ngIf="!shortened">
+      Note: Some courses don't have grade or enrollment information
+    </mat-card-subtitle>
+    <mat-card-content>
+      <mat-button-toggle-group name="fontStyle" aria-label="Font Style">
+        <mat-button-toggle value="all" (click)="changeSpecialization('all')">{{ shortened ? 'ALL' : 'All' }}</mat-button-toggle>
+        <mat-button-toggle value="cpr" (click)="changeSpecialization('cpr')">{{ shortened ? 'CPR' : 'Computational Perception and Robotics'}}</mat-button-toggle>
+        <mat-button-toggle value="cs" (click)="changeSpecialization('cs')">{{ shortened ? 'CS' : 'Computing Systems'}}</mat-button-toggle>
+        <mat-button-toggle value="ii" (click)="changeSpecialization('ii')">{{ shortened ? 'II' : 'Interactive Intelligence'}}</mat-button-toggle>
+        <mat-button-toggle value="ml" (click)="changeSpecialization('ml')">{{ shortened ? 'ML' : 'Machine Learning'}}</mat-button-toggle>
+      </mat-button-toggle-group>
+    </mat-card-content>
+    <!-- <mat-card-actions>
+      <button mat-raised-button [color]="percent ? '' : 'accent'" (click)="percent = !percent">Counts</button>
+      <button mat-raised-button [color]="percent ? 'accent' : ''" (click)="percent = !percent">Percentage</button>
+      <span *ngIf="percent">&nbsp;&nbsp;Percentages are based off of A-F grade counts.</span>
+    </mat-card-actions> -->
+  </mat-card>
+  <div class="courses-header">
+    <mat-form-field>
+      <input matInput (keyup)="applyFilter($event.target.value)" placeholder="Filter">
+    </mat-form-field>
+  </div>
+
+  <div class="courses-container mat-elevation-z8">
+
+    <table mat-table [dataSource]="dataSource" matSort>
+
+      <!-- combined Column -->
+      <ng-container matColumnDef="combined">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Course</th>
+        <td mat-cell *matCellDef="let row">
+          <a mat-button [routerLink]="['/courses', row.id]">{{row.combined}}</a>
+        </td>
+      </ng-container>
+
+      <!-- enrolled Column -->
+      <ng-container matColumnDef="enrolled">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Enrolled</th>
+        <td mat-cell *matCellDef="let row"> {{row.enrolled | number}} </td>
+      </ng-container>
+
+      <!-- reviews Column -->
+      <ng-container matColumnDef="numReviews">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Reviews</th>
+        <td mat-cell *matCellDef="let row"> {{row.numReviews | number}} </td>
+      </ng-container>
+
+      <!-- work Column -->
+      <ng-container matColumnDef="work">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Workload</th>
+        <td mat-cell *matCellDef="let row"> {{row.work | number}} </td>
+      </ng-container>
+
+      <!-- difficulty Column -->
+      <ng-container matColumnDef="difficulty">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Difficulty</th>
+        <td mat-cell *matCellDef="let row"> {{row.difficulty | number}} </td>
+      </ng-container>
+
+      <!-- rating Column -->
+      <ng-container matColumnDef="rating">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Rating</th>
+        <td mat-cell *matCellDef="let row"> {{row.rating| number}} </td>
+      </ng-container>
+
+      <!-- ab Column -->
+      <ng-container matColumnDef="ab">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>&ge; B</th>
+        <td mat-cell *matCellDef="let row"> {{row.ab | number}} </td>
+      </ng-container>
+
+      <!-- cdf Column -->
+      <ng-container matColumnDef="cdf">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>&le; C</th>
+        <td mat-cell *matCellDef="let row"> {{row.cdf | number}} </td>
+      </ng-container>
+
+      <!-- withdrew Column -->
+      <ng-container matColumnDef="withdrew">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Withdrew</th>
+        <td mat-cell *matCellDef="let row"> {{row.withdrew | number}} </td>
+      </ng-container>
+
+      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+      <tr mat-row *matRowDef="let row; columns: displayedColumns;">
+      </tr>
+    </table>
+
+    <mat-paginator [pageSizeOptions]="shortened ? [1000] : [10, 25, 100]"></mat-paginator>
+  </div>
