--- src/app/courses/courses.component.scss
@@ -1,22 +1,0 @@
-.course-text {
-  padding-top: .5px;
-  padding-bottom: .5px;
-}
-
-.courses-header {
-  min-height: 64px;
-  padding: 8px 24px 0;
-}
-
-.mat-form-field {
-  font-size: 14px;
-  width: 100%;
-}
-
-.courses-container {
-  display: flex;
-  flex-direction: column;
-  min-width: 300px;
-  max-height: 500px;
-  overflow: auto;
-}
