+++ src/styles.scss
@@ -58,3 +58,17 @@ oms-nav {
   justify-content: center;
   width: 200px;
 }
+
+.coursenav {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 300px;
+}
+
+.courseinfo {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 400px;
+}
